photocalpro/requirements.txt
beautifulsoup4==4.10.0
attrs==21.4.0
bleach==4.1.0
blinker==1.4
decorator==5.1.1
easygui==0.98.2
entrypoints==0.4
executing==0.8.3
f==0.0.1
fit==0.4.1
Flask==2.1.3
flatbuffers==2.0
fonttools==4.30.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.44.0
grpcio-reflection==1.44.0
grpcio-tools==1.44.0
h5==0.5.2
h5py==3.6.0
heatmapz==0.0.4
httplib2==0.20.4
idna==2.8
imageio==2.16.1
importlib-metadata==4.11.3
int==3.2.0
iopath==0.1.10
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
Js2Py==0.71
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.4
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
label==0.1.1
layoutparser==0.3.4
libclang==13.0.0
littleutils==0.2.2
lxml==4.8.0
map==1.3.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
object-detection==0.0.3
opencv-python==4.5.5.64
opt-einsum==3.3.0
Optimizers==2.1
outdated==0.2.1
packaging==21.3
pandas==1.4.1
pandas-flavor==0.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pdf2image==1.16.0
pdfminer.six==20220524
pdfplumber==0.7.4
pickleshare==0.7.5
Pillow==9.2.0
pingouin==0.5.1
pyreadline==2.1
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pyttsx3==2.90
QtPy==2.1.0
regex==2022.3.15
# requests==2.21.0
requests-oauthlib==1.3.1
roman==3.3
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
tf-estimator-nightly==2.8.0.dev2021122109
tflearn==0.5.0
widgetsnbextension==3.5.2
xarray==2022.3.0